feat: allow custom profiles in vizdeps
Currently the only supported profiles in the `vizdeps` command are
`:user` and `:dev`. `leiningen` allows arbitrary profiles to be created
and to have custom dependencies (and other fields) defined for those
profiles.

This commit
-----------
Adds a `-P | --profiles` parameter, expecting a comma-delimited string
of profile names. These profiles will then be used in dependency
calculation. The `-d | --dev` flag is removed as its functionality is
superseded by `-P | --profiles`.

Note that this is a _BREAKING_ change, due to the removal of the `-d`
flag.

Relates to
----------
walmartlabs/vizdeps#16